RDecl, RTypeDecl, RFuncDecl, RBodyDeclOuter, RTypeDeclOuter, RTopLevelDeclOuter 구현
RDecl, RTypeDecl과 RDecl, RFuncDecl합치기

+++ src/IR0/Public/IR0/RIdentifier.h
@@ -15,7 +15,7 @@ using RTypePtr = std::shared_ptr<RType>;
 struct RIdentifier
 {
     RName name;
-    int typeParamCount;
+    size_t typeParamCount;
     std::vector<RTypePtr> paramIds;
 
     bool operator==(const RIdentifier& other) const = default;
